Conventional view rendering (could be better...)
I was hoping to get automatic rendering if no call to render or redirect had been used (like in Rails) but it seems really hairy to get that sort of dynamic thing going in Crystal.

Will probably end up requiring a manual render call and build classes on the fly if needed (instead of the run view_compiler thing).